+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## 1.3.0 --- Linearization Effort (WIP)
+
+- Linear parsimony optimization
+
 ## 1.2.1 --- Reporting Bias (WIP)
 
-- Export `analysis` and `readwrite` module.
 - Report feasible communities, even if optimization fails
 - Add function to get `supplier` and `community` frozensets.
 - Update crossfeed to report as frozensets.